refactor: simplify dashboard panels
