+++ plugins/codex-autoresearch/dashboard/src/components/trend/ExperimentModal.tsx
@@ -1,0 +1,170 @@
+import { useEffect, useRef } from "react";
+import type { KeyboardEvent } from "react";
+import { formatImprovement, formatMetric } from "../../model";
+import type { DashboardReadout, RunMetricBreakdown } from "../../types";
+import { focusFallback, getFocusableDialogElements } from "./focus";
+import {
+  formatChartAxisValue,
+  formatMemoryValue,
+  formatWeightedScoreValue,
+  type ChartDatum,
+  type ValueMode,
+} from "./shared";
+
+export function ExperimentModal({
+  point,
+  valueMode,
+  readout,
+  restoreFocusSelector,
+  onClose,
+}: {
+  point: ChartDatum;
+  valueMode: ValueMode;
+  readout: DashboardReadout;
+  restoreFocusSelector: string;
+  onClose: () => void;
+}) {
+  const breakdown = point.breakdown;
+  const dialogRef = useRef<HTMLElement | null>(null);
+  const closeRef = useRef<HTMLButtonElement | null>(null);
+  useEffect(() => {
+    closeRef.current?.focus();
+  }, []);
+  const requestClose = () => {
+    onClose();
+    window.setTimeout(() => {
+      focusFallback(restoreFocusSelector)?.focus();
+    }, 600);
+  };
+  const onDialogKeyDown = (event: KeyboardEvent<HTMLElement>) => {
+    if (event.key === "Escape") {
+      event.preventDefault();
+      requestClose();
+      return;
+    }
+    if (event.key !== "Tab") return;
+    const focusable = getFocusableDialogElements(dialogRef.current);
+    if (!focusable.length) return;
+    const first = focusable[0];
+    const last = focusable[focusable.length - 1];
+    if (event.shiftKey && document.activeElement === first) {
+      event.preventDefault();
+      last.focus();
+    } else if (!event.shiftKey && document.activeElement === last) {
+      event.preventDefault();
+      first.focus();
+    }
+  };
+  return (
+    <div className="modal-backdrop" role="presentation" onMouseDown={requestClose}>
+      <section
+        ref={dialogRef}
+        className="experiment-modal"
+        role="dialog"
+        aria-modal="true"
+        aria-labelledby="experiment-modal-title"
+        onMouseDown={(event) => event.stopPropagation()}
+        onKeyDown={onDialogKeyDown}
+      >
+        <button
+          ref={closeRef}
+          className="modal-close"
+          type="button"
+          aria-label="Close experiment details"
+          onClick={requestClose}
+        >
+          {"\u00d7"}
+        </button>
+        <p className="eyebrow">
+          {point.statusLabel} / {point.timestampLabel || "no timestamp"}
+        </p>
+        <h2 id="experiment-modal-title">Run #{point.runNumber}</h2>
+        <div className="experiment-metrics">
+          <div>
+            <span>{readout.metricDefinition.valueLabel}</span>
+            <strong>{point.metricDisplay}</strong>
+          </div>
+          <div>
+            <span>Improvement</span>
+            <strong>{formatImprovement(breakdown?.improvement ?? null)}</strong>
+          </div>
+          <div>
+            <span>{readout.metricDefinition.percentLabel}</span>
+            <strong>
+              {formatChartAxisValue(
+                valueMode === "percent" ? point.chartPercent : point.metric,
+                valueMode,
+                readout,
+              )}
+            </strong>
+          </div>
+        </div>
+        <dl className="experiment-detail-list">
+          <div>
+            <dt>Experiment</dt>
+            <dd>{point.description}</dd>
+          </div>
+          {point.heldMetric && (
+            <div>
+              <dt>Chart placement</dt>
+              <dd>Crash plotted at the nearest successful metric level.</dd>
+            </div>
+          )}
+          {point.hypothesis && (
+            <div>
+              <dt>What was tried</dt>
+              <dd>{point.hypothesis}</dd>
+            </div>
+          )}
+          {point.evidence && (
+            <div>
+              <dt>Evidence</dt>
+              <dd>{point.evidence}</dd>
+            </div>
+          )}
+          <WeightedExperimentMetrics readout={readout} breakdown={breakdown} />
+          {point.rollbackReason && (
+            <div>
+              <dt>Rollback reason</dt>
+              <dd>{point.rollbackReason}</dd>
+            </div>
+          )}
+          {point.nextActionHint && (
+            <div>
+              <dt>Next action</dt>
+              <dd>{point.nextActionHint}</dd>
+            </div>
+          )}
+        </dl>
+      </section>
+    </div>
+  );
+}
+
+function WeightedExperimentMetrics({
+  readout,
+  breakdown,
+}: {
+  readout: DashboardReadout;
+  breakdown: RunMetricBreakdown | null;
+}) {
+  if (!breakdown || readout.metricDefinition.mode !== "weighted_cost") return null;
+  return (
+    <>
+      <div>
+        <dt>Time</dt>
+        <dd>
+          {formatMetric(breakdown.timeValue, "s")} / score{" "}
+          {formatWeightedScoreValue(breakdown.timeScore, readout)}
+        </dd>
+      </div>
+      <div>
+        <dt>Memory</dt>
+        <dd>
+          {formatMemoryValue(breakdown.memoryValue)} / score{" "}
+          {formatWeightedScoreValue(breakdown.memoryScore, readout)}
+        </dd>
+      </div>
+    </>
+  );
+}
