+++ plugins/codex-autoresearch/dashboard/src/styles.css
@@ -2379,14 +2379,9 @@ p {
 }
 
 .ledger-row .status-pill {
-  display: inline-flex;
-  align-items: center;
   border: 1px solid var(--line);
-  border-radius: 999px;
   padding: 5px 8px;
   font-size: 11px;
-  font-weight: 900;
-  text-transform: uppercase;
 }
 
 .ledger-row .status-pill.keep {
